+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.1.7]
+### Added
+- add `-j<N>` to build/task, pass to ninja
 ### Fixed
 - fix release workflow